+++ lib/engine/game/g_1873/round/auction.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require_relative '../../../round/auction'
+
+module Engine
+  module Game
+    module G1873
+      module Round
+        class Auction < Engine::Round::Auction
+          def self.short_name
+            'AR'
+          end
+        end
+      end
+    end
+  end
+end
