--- lib/engine/round/g_1873/auction.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../auction'
-
-module Engine
-  module Round
-    module G1873
-      class Auction < Auction
-        def self.short_name
-          'AR'
-        end
-      end
-    end
-  end
-end
